Move getVersion() to a Trait for easy reuse in TeamQuestionBankController.
